% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIRTL.R
\name{get_well_subset}
\alias{get_well_subset}
\title{Given a row and column range, return a vector that includes all wells in the range}
\usage{
get_well_subset(row_range = 1:16, col_range = 1:24)
}
\arguments{
\item{row_range}{Range of rows on the plate}

\item{col_range}{Range of columns on the plate}
}
\value{
Vector of well IDs
}
\description{
Given a row and column range, return a vector that includes all wells in the range
}
