% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIRTL.R
\name{run_single_point_analysis_sub_gpu}
\alias{run_single_point_analysis_sub_gpu}
\title{Run a chain pairing analysis}
\usage{
run_single_point_analysis_sub_gpu(
  folder_path,
  outdir = getwd(),
  prefix = "tmp",
  well_filter_thres = 0.5,
  min_reads = 0,
  min_wells = 2,
  well_pos = 3,
  wellset1 = get_well_subset(1:16, 1:24),
  compute = T,
  backend = "numpy",
  file_pattern = NULL
)
}
\arguments{
\item{outdir}{Path where output should be saved}

\item{prefix}{Prefix added to the output files}

\item{foler_path}{Path to the directory containing the clone files}
}
\value{
Data frame with clone pairings and MAD-HYPE / TSHELL scores.
}
\description{
Run a chain pairing analysis
}
